Add signature to highlight CAPE family detection
Adding in a signature to highlight family detection for CAPE. While this is visually presented already in several places this is merely to tie it to a signature so that it improves the score further.

I have tested and it is working but please have an check over the signature for the logic just in case I am missing something.

(cherry picked from commit ec600971b6636291786cff57541f1ef263749059)